refactor(options_parser): ♻️ | change frequency type to unsigned int
updates frequency variable type in options struct to unsigned int for better clarity and potential range improvement

+++ include/server/options_parser/options.h
@@ -19,7 +19,7 @@ typedef struct server_options_s {
     size_t height;
     char **teams;
     size_t clients_nb;
-    size_t frequency;
+    unsigned int frequency;
     bool help;
     bool debug;
     bool error;
